[prj1] complete pass_one

+++ projects/1/src/translate_utils.c
@@ -65,10 +65,10 @@ int is_valid_label(const char *str)
    function returns 0 if the conversion proceeded without errors, or -1 if an 
    error occurred.
  */
-int translate_num(long int *output, const char *str, long int lower_bound,
-                  long int upper_bound)
+int translate_num(int *output, const char *str, int lower_bound,
+                  int upper_bound)
 {
-    long res;
+    int res;
     char *pEnd;
 
     if (!str || !output)
